+++ bika/lims/browser/worksheet/views/add_analyses.py
@@ -52,7 +52,7 @@ class AddAnalysesView(BikaListingView):
 
         # initial review state for first form display of the worksheet
         # add_analyses search view - first batch of analyses, latest first.
-        self.show_sort_column = False
+
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 50