+++ bika/lims/controlpanel/bika_analysiscategories.py
@@ -51,7 +51,6 @@ class AnalysisCategoriesView(BikaListingView):
             "/++resource++bika.lims.images/category_big.png"
         )
 
-        self.show_sort_column = False
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 25