+++ bika/lims/controlpanel/bika_analysisspecs.py
@@ -13,10 +13,8 @@ from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import IAnalysisSpecs
 from bika.lims.utils import get_link
-from plone.app.content.browser.interfaces import IFolderContentsView
 from plone.app.folder.folder import ATFolder
 from plone.app.folder.folder import ATFolderSchema
-from plone.app.layout.globals.interfaces import IViewView
 from Products.Archetypes import atapi
 from Products.ATContentTypes.content import schemata
 from zope.interface.declarations import implements
@@ -54,7 +52,6 @@ class AnalysisSpecsView(BikaListingView):
             "/++resource++bika.lims.images/analysisspec_big.png"
         )
 
-        self.show_sort_column = False
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 25