+++ bika/lims/controlpanel/bika_departments.py
@@ -14,10 +14,8 @@ from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import IDepartments
 from bika.lims.utils import get_email_link
 from bika.lims.utils import get_link
-from plone.app.content.browser.interfaces import IFolderContentsView
 from plone.app.folder.folder import ATFolder
 from plone.app.folder.folder import ATFolderSchema
-from plone.app.layout.globals.interfaces import IViewView
 from Products.Archetypes import atapi
 from Products.Archetypes.utils import DisplayList
 from Products.ATContentTypes.content import schemata
@@ -28,7 +26,6 @@ from zope.interface.declarations import implements
 
 
 class DepartmentsView(BikaListingView):
-    implements(IFolderContentsView, IViewView)
 
     def __init__(self, context, request):
         super(DepartmentsView, self).__init__(context, request)
@@ -54,7 +51,6 @@ class DepartmentsView(BikaListingView):
             "/++resource++bika.lims.images/department_big.png"
         )
 
-        self.show_sort_column = False
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 25