+++ bika/lims/controlpanel/bika_labproducts.py
@@ -5,26 +5,17 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from AccessControl import ClassSecurityInfo
 from Products.ATContentTypes.content import schemata
 from Products.Archetypes import atapi
-from Products.Archetypes.ArchetypeTool import registerType
-from Products.CMFCore import permissions
-from Products.CMFCore.utils import getToolByName
-from bika.lims.browser import BrowserView
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims import bikaMessageFactory as _
-from bika.lims.utils import t
-from plone.app.layout.globals.interfaces import IViewView
-from bika.lims.content.bikaschema import BikaFolderSchema
-from plone.app.content.browser.interfaces import IFolderContentsView
 from plone.app.folder.folder import ATFolder, ATFolderSchema
 from zope.interface.declarations import implements
 from bika.lims.interfaces import ILabProducts
 
+
 class LabProductsView(BikaListingView):
-    implements(IFolderContentsView, IViewView)
 
     def __init__(self, context, request):
         super(LabProductsView, self).__init__(context, request)
@@ -37,7 +28,7 @@ class LabProductsView(BikaListingView):
         self.title = self.context.translate(_("Lab Products"))
         self.icon = self.portal_url + "/++resource++bika.lims.images/product_big.png"
         self.description = ""
-        self.show_sort_column = False
+
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 25
@@ -93,6 +84,12 @@ class LabProductsView(BikaListingView):
                          'VATAmount',
                          'TotalPrice']},
         ]
+
+    def before_render(self):
+        """Before template render hook
+        """
+        # Don't allow any context actions
+        self.request.set("disable_border", 1)
 
     def folderitems(self):
         items = BikaListingView.folderitems(self)