+++ bika/lims/controlpanel/bika_manufacturers.py
@@ -5,30 +5,22 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from AccessControl import ClassSecurityInfo
 from Products.ATContentTypes.content import schemata
 from Products.Archetypes import atapi
-from Products.Archetypes.ArchetypeTool import registerType
-from Products.CMFCore import permissions
-from Products.CMFCore.utils import getToolByName
-from bika.lims.browser import BrowserView
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims import bikaMessageFactory as _
-from bika.lims.utils import t
-from plone.app.layout.globals.interfaces import IViewView
-from bika.lims.content.bikaschema import BikaFolderSchema
-from plone.app.content.browser.interfaces import IFolderContentsView
 from plone.app.folder.folder import ATFolder, ATFolderSchema
 from bika.lims.interfaces import IManufacturers
 from zope.interface.declarations import implements
 
+
 class ManufacturersView(BikaListingView):
-    implements(IFolderContentsView, IViewView)
+
     def __init__(self, context, request):
         super(ManufacturersView, self).__init__(context, request)
         self.catalog = 'bika_setup_catalog'
-        self.contentFilter = {'portal_type': 'Manufacturer', 
+        self.contentFilter = {'portal_type': 'Manufacturer',
                               'sort_on': 'sortable_title'}
         self.context_actions = {_('Add'):
                                 {'url': 'createObject?type_name=Manufacturer',
@@ -36,7 +28,7 @@ class ManufacturersView(BikaListingView):
         self.title = self.context.translate(_("Manufacturers"))
         self.icon = "++resource++bika.lims.images/manufacturer_big.png"
         self.description = ""
-        self.show_sort_column = False
+
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 25
@@ -66,6 +58,12 @@ class ManufacturersView(BikaListingView):
              'columns': ['Title', 'Description']},
         ]
 
+    def before_render(self):
+        """Before template render hook
+        """
+        # Don't allow any context actions
+        self.request.set("disable_border", 1)
+
     def folderitems(self):
         items = BikaListingView.folderitems(self)
         for x in range(len(items)):
@@ -76,7 +74,10 @@ class ManufacturersView(BikaListingView):
                  (items[x]['url'], items[x]['Title'])
         return items
 
+
 schema = ATFolderSchema.copy()
+
+
 class Manufacturers(ATFolder):
     implements(IManufacturers)
     displayContentsTab = False