+++ bika/lims/controlpanel/bika_setupitems.py
@@ -5,24 +5,11 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from AccessControl.SecurityInfo import ClassSecurityInfo
-from Products.ATContentTypes.content import schemata
-from Products.Archetypes import atapi
 from bika.lims import bikaMessageFactory as _
-from bika.lims.utils import t
-from Products.CMFCore.utils import getToolByName
 from bika.lims.browser.bika_listing import BikaListingView
-from bika.lims.config import PROJECTNAME
-from bika.lims.interfaces import IAnalysisCategories
-from plone.app.content.browser.interfaces import IFolderContentsView
-from plone.app.layout.globals.interfaces import IViewView
-from plone.app.folder.folder import ATFolderSchema, ATFolder
-from zope.interface.declarations import implements
-from zope.interface import alsoProvides
 
 
 class BikaSetupItemsView(BikaListingView):
-    implements(IFolderContentsView, IViewView)
 
     def __init__(self, context, request, typename, iconfilename):
         super(BikaSetupItemsView, self).__init__(context, request)
@@ -67,3 +54,9 @@ class BikaSetupItemsView(BikaListingView):
              'contentFilter': {},
              'columns': ['Title', 'Description']},
         ]
+
+    def before_render(self):
+        """Before template render hook
+        """
+        # Don't allow any context actions
+        self.request.set("disable_border", 1)