+++ bika/lims/controlpanel/bika_worksheettemplates.py
@@ -50,7 +50,6 @@ class WorksheetTemplatesView(BikaListingView):
             "/++resource++bika.lims.images/worksheettemplate_big.png"
         )
 
-        self.show_sort_column = False
         self.show_select_row = False
         self.show_select_column = True